refactor: move cast details to detail screen
