+++ domain/src/main/java/com/amsterdam/domain/useCase/details/GetMovieCastUseCase.kt
@@ -1,4 +1,4 @@
-package com.amsterdam.domain.useCase.common
+package com.amsterdam.domain.useCase.details
 
 import com.amsterdam.domain.repository.MovieRepository
 import com.amsterdam.entity.Actor
