+++ domain/src/test/java/com/amsterdam/domain/useCase/GetMovieCastUseCaseTest.kt
@@ -2,7 +2,7 @@ package com.amsterdam.domain.useCase
 
 import com.amsterdam.domain.exceptions.AflamiException
 import com.amsterdam.domain.repository.MovieRepository
-import com.amsterdam.domain.useCase.common.GetMovieCastUseCase
+import com.amsterdam.domain.useCase.details.GetMovieCastUseCase
 import com.amsterdam.entity.Actor
 import com.amsterdam.entity.Gender
 import com.google.common.truth.Truth.assertThat
